+++ lib/cannery/ammo/ammo_group.ex
@@ -30,6 +30,7 @@ defmodule Cannery.Ammo.AmmoGroup do
     field :notes, :string
     field :price_paid, :float
     field :staged, :boolean, default: false
+    field :purchased_on, :date
 
     belongs_to :ammo_type, AmmoType
     belongs_to :container, Container
@@ -46,6 +47,7 @@ defmodule Cannery.Ammo.AmmoGroup do
           notes: String.t() | nil,
           price_paid: float() | nil,
           staged: boolean(),
+          purchased_on: Date.t(),
           ammo_type: AmmoType.t() | nil,
           ammo_type_id: AmmoType.id(),
           container: Container.t() | nil,
@@ -79,9 +81,9 @@ defmodule Cannery.Ammo.AmmoGroup do
     |> change(ammo_type_id: ammo_type_id)
     |> change(user_id: user_id)
     |> change(container_id: container_id)
-    |> cast(attrs, [:count, :price_paid, :notes, :staged])
+    |> cast(attrs, [:count, :price_paid, :notes, :staged, :purchased_on])
     |> validate_number(:count, greater_than: 0)
-    |> validate_required([:count, :staged, :ammo_type_id, :container_id, :user_id])
+    |> validate_required([:count, :staged, :purchased_on, :ammo_type_id, :container_id, :user_id])
   end
 
   @doc """
@@ -99,10 +101,10 @@ defmodule Cannery.Ammo.AmmoGroup do
           Changeset.t(t() | new_ammo_group())
   def update_changeset(ammo_group, attrs, user) do
     ammo_group
-    |> cast(attrs, [:count, :price_paid, :notes, :staged, :container_id])
+    |> cast(attrs, [:count, :price_paid, :notes, :staged, :purchased_on, :container_id])
     |> validate_number(:count, greater_than_or_equal_to: 0)
     |> validate_container_id(user)
-    |> validate_required([:count, :staged, :container_id])
+    |> validate_required([:count, :staged, :purchased_on, :container_id])
   end
 
   defp validate_container_id(changeset, user) do